pybricks.experimental: Refactor flash code.
Make it somewhat more generic so we can experiment with reading/saving
a main.py user script from main.c

Also enlarge buffers.

Also add mkdir command to create /_pybricks if it doesn't exist.
